# docs/swagger.yaml
definitions:
  handlers.AuthDTO:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  handlers.Credentials:
    properties:
      accessToken:
        type: string
      refreshTokent:
        type: string
    type: object
  handlers.UserDTO:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Аутентификация пользователя с возвратом access и refresh токенов
      parameters:
      - description: Данные для авторизации
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.AuthDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.Credentials'
        "400":
          description: Ошибка валидации
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Неверный логин или пароль
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Авторизация пользователя
      tags:
      - auth
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Используется для получения нового access токена по refresh токену
      parameters:
      - description: Refresh токен
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.Credentials'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.Credentials'
        "400":
          description: Ошибка валидации
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Неверный refresh токен
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновление access токена
      tags:
      - auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Создание нового пользователя с указанными данными
      parameters:
      - description: Данные для регистрации
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.UserDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Пользователь успешно зарегистрирован
          schema:
            type: string
        "400":
          description: Ошибка валидации
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Пользователь уже существует
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Регистрация нового пользователя
      tags:
      - auth
swagger: "2.0"